# Only run ts checker and eslint plugins in webpack dev mode